refactor: consolidate standards constants into barrel export
Add src/lib/constants/index.ts re-exporting wcag, section508, and en301549
constants. Update all direct per-file imports to go through the barrel.
