+++ src/components/issues/section508-selector.tsx
@@ -3,7 +3,7 @@ import { X } from 'lucide-react';
 import {
   SECTION_508_CRITERION_CODES,
   SECTION_508_CRITERION_NAMES,
-} from '@/lib/constants/section508';
+} from '@/lib/constants';
 import { Button } from '@/components/ui/button';
 
 interface Section508SelectorProps {
